Add README.md
Add the primary README.md and supporting image. This probably should not
be the final state of this document, for example. File Overview should
probably not exist and instead they should be added as comments.
Additionally, there should probably be a full set of super simple steps
for the user to get up and running.